// backend/generated/prisma/models/WatchProgress.ts
export type WatchProgressModel = runtime.Types.Result.DefaultSelection<Prisma.$WatchProgressPayload>
// ...
export type $WatchProgressPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "WatchProgress"
  objects: {
    usuario: Prisma.$UsuarioPayload<ExtArgs>
    episodio: Prisma.$EpisodioPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    usuarioId: number
    episodioId: number
    assistido: boolean
    segundosAssistidos: number
    porcentagem: number
    atualizadoEm: Date
  }, ExtArgs["result"]["watchProgress"]>
  composites: {}
}